requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
flake8==7.1.2
flake8-isort==6.1.2
inflection==0.5.1
iniconfig==2.0.0
isort==6.0.0
mccabe==0.7.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
